Revert "Added test for recursion check in python implementation"
This reverts commit 9a262f5e7b499037e771eb14044dde94b88a7bba.

According to https://github.com/wekan/wekan/issues/2545,
this file contains the WannaCry virus.

I understand having a test that ensures a virus is properly
scanned is interesting, but I think for the sake of the users,
we should not redistribute such file in a useful tool like this
one.